reports-service/requirements.txt
# ============================================================================
# BionicPRO Reports Service Dependencies
# ============================================================================

# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# ClickHouse client
clickhouse-driver==0.2.6

# Redis for caching
redis==5.0.1

# JWT authentication
python-jose[cryptography]==3.3.0

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client for auth service
httpx==0.26.0

# Date handling
python-dateutil==2.8.2
